+++ P2PLearningAPI/Repository/PostRepository.cs
@@ -8,10 +8,14 @@ namespace P2PLearningAPI.Repository
     public class PostRepository : IPostInterface
     {
         private readonly P2PLearningDbContext _context;
+        private readonly ITokenService _tokenService;
+        private readonly ILogger<PostRepository> _logger;
 
-        public PostRepository(P2PLearningDbContext context)
+        public PostRepository(P2PLearningDbContext context, ITokenService tokenService, ILogger<PostRepository> logger)
         {
             _context = context;
+            _tokenService = tokenService;
+            _logger = logger;
         }
 
         // Get all posts
@@ -21,13 +25,13 @@ namespace P2PLearningAPI.Repository
         }
 
         // Get a single post by ID
-        public Post GetPost(long id)
+        public Post? GetPost(long id)
         {
             return _context.Posts.FirstOrDefault(p => p.Id == id)!;
         }
 
         // Get posts by user ID
-        public ICollection<Post> GetPostsByUser(long userId)
+        public ICollection<Post> GetPostsByUser(string userId)
         {
             return _context.Posts.Where(p => p.UserID == userId).ToList();
         }
@@ -39,39 +43,31 @@ namespace P2PLearningAPI.Repository
         }
 
         // Create a new post
-        public Post CreatePost(PostDTO postDTO, PostType postType)
+        public Post CreatePost(PostDTO postDTO, PostType postType, string token)
         {
-            if (postDTO == null) 
-                throw new ArgumentNullException(nameof(postDTO));
-            Post newPost;
-            switch (postType)
-            {
-                case PostType.Question:
-                    if(postDTO.Discussion == null)
-                        throw new ArgumentNullException(nameof(postDTO.Discussion));
-                        newPost = new Question(postDTO.Title, postDTO.Content,postDTO.PostedBy, postDTO.Discussion);
-                    break;
-                case PostType.Answer:
-                    if (postDTO.Question == null)
-                        throw new ArgumentException(nameof(postDTO.Question));
-                    newPost = new Answer(postDTO.Title, postDTO.Content, postDTO.PostedBy, postDTO.Question);
-                    break;
-                default:
-                    throw new ArgumentException("Invalid post type", nameof(postType));
-            }
+            var (userId, _, _) = _tokenService.DecodeToken(token);
+            if (userId != postDTO.PostedBy.Id)
+                throw new UnauthorizedAccessException();
+
+            var newPost = CreatePostFromDTO(postDTO, postType);
             _context.Posts.Add(newPost);
-            if(Save())
-                return newPost;
-            throw new InvalidOperationException("Failed to save the Post to the database.");
+
+            if (!Save())
+                throw new InvalidOperationException("Failed to save the post");
+
+            return newPost;
         }
 
         // Update an existing post
-        public Post UpdatePost(Post post)
+        public Post UpdatePost(Post post, string token)
         {
             if(post == null)
                 throw new ArgumentNullException(nameof(post));
             if (!CheckPostExist(post.Id))
                 throw new InvalidOperationException("Post doesn't exist");
+            (string userId, _, _) = _tokenService.DecodeToken(token);
+            if (userId != post.UserID)
+                throw new UnauthorizedAccessException("User is not authorized to update this post.");
             _context.Posts.Update(post);
             if(Save())
                 return post;
@@ -79,24 +75,30 @@ namespace P2PLearningAPI.Repository
         }
 
         // Close a post by setting IsClosed to true
-        public bool ClosePost(long id)
+        public bool ClosePost(long id, string token)
         {
             var post = GetPost(id);
+            var (userId, _, _) = _tokenService.DecodeToken(token);
             if (post == null)
-                return false;
-
+                throw new InvalidOperationException("Post doesn't exist");
+            if (userId != post.UserID)
+                throw new UnauthorizedAccessException("User is not authorized to close this post.");
             post.IsClosed = true;
             return Save();
             
         }
 
         // Reopen a post by setting IsClosed to false
-        public bool ReopenPost(long id)
+        public bool ReopenPost(long id, string token)
         {
             var post = GetPost(id);
             if (post == null)
                 return false;
-
+            var (userId, _, _) = _tokenService.DecodeToken(token);
+            if (userId != post.UserID)
+                throw new UnauthorizedAccessException("User is not authorized to reopen this post.");
+            if (!post.IsClosed)
+                throw new InvalidOperationException("Post is already open");
             post.IsClosed = false;
             return Save();
         }
@@ -121,12 +123,14 @@ namespace P2PLearningAPI.Repository
         }
 
         // Delete a post by ID
-        public bool DeletePost(long id)
+        public bool DeletePost(long id, string token)
         {
             var post = GetPost(id);
             if (post == null)
-                return false;
-
+                throw new InvalidOperationException("Post doesn't exist");
+            var (userId, _, userType) = _tokenService.DecodeToken(token);
+            if (userId != post.UserID && userType != "Administrator")
+                throw new UnauthorizedAccessException("User is not authorized to delete this post.");
             _context.Posts.Remove(post);
             return Save();
         }
@@ -135,6 +139,17 @@ namespace P2PLearningAPI.Repository
         public bool Save()
         {
             return _context.SaveChanges() > 0;
+        }
+
+        public Post CreatePostFromDTO(PostDTO postDTO, PostType postType)
+        {
+            return postType switch
+            {
+                PostType.Question => new Question(postDTO.Title, postDTO.Content, postDTO.PostedBy, postDTO.Discussion ?? throw new ArgumentNullException(nameof(postDTO.Discussion))),
+                PostType.Answer => new Answer(postDTO.Title, postDTO.Content, postDTO.PostedBy, postDTO.Question ?? throw new ArgumentNullException(nameof(postDTO.Question))),
+                PostType.Reply => new Answer(postDTO.Title, postDTO.Content, postDTO.PostedBy, postDTO.Answer ?? throw new ArgumentNullException(nameof(postDTO.Answer))),
+                _ => throw new ArgumentException("Invalid post type", nameof(postType))
+            };
         }
     }
 }
